requirements.txt
apache-airflow==2.7.3
requests==2.31.0
pandas==2.0.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
